Features/requirements.txt
allure-behave==2.12.0
allure-combine==1.0.8
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
behave==1.2.6
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.0.4
Flask==2.2.2
h11==0.14.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
outcome==1.2.0
parse==1.19.0
parse-type==0.6.0
Pillow==9.3.0
pluggy==1.0.0
pycparser==2.21
PySocks==1.7.1
PyWavelets==1.4.1
scipy==1.9.3
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.10.0
